requirements.txt
Django==4.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
python-decouple==3.8
# psycopg2-binary==2.9.6 binary версия для мака и линукса,
